pyproject.toml
[tool.poetry]
name = "api-testing-agent"
version = "0.1.0"
description = "AI chatbot for OpenAPI-based REST API testing"
authors = ["Huy Anh <huynhnguyenhuyanh.work@gmail.com>"]
readme = "README.md"
packages = [{ include = "api_testing_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.0.1"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
langchain = "^1.2.15"
langchain-openai = "^1.2.1"
langgraph = "^1.1.9"
langgraph-checkpoint-sqlite = "^3.0.3"
langchain-anthropic = "^1.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.7"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"